Replace Http facade usage with GuzzleHTTP client for Laravel versions prior to 7.x
This pull request replaces the usage of the HTTP facade with direct usage of the GuzzleHTTP client in Laravel versions before 7.x.

In Laravel versions before 7.x, the HTTP facade is not available, and developers typically utilize the GuzzleHTTP client directly for making HTTP requests. This pull request ensures compatibility with older Laravel versions by replacing the HTTP facade usage with the GuzzleHTTP client.